backend/requirements.txt
fastapi
uvicorn
passlib==1.7.4
bcrypt==3.2.0
python-jose[cryptography]
psycopg2-binary
httpx